# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  handler.LoginData:
    properties:
      access_expires_at:
        type: string
      access_token:
        type: string
      token_type:
        type: string
    type: object
  handler.LoginReq:
    properties:
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 64
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  handler.LoginResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.LoginData'
      msg:
        type: string
    type: object
  handler.LogoutResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  handler.RegisterResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/models.User'
      msg:
        type: string
    type: object
...
  handler.UpdateUserResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      timezone:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: 管理API
  title: ToDoList API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码进行身份验证，获取JWT token
      parameters:
      - description: 登录请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回token
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 登录失败
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 用户登录
  /logout:
    post:
      description: 需要有效的JWT token认证，无需请求体
      produces:
      - application/json
      responses:
        "200":
          description: 退出登录成功
          schema:
            $ref: '#/definitions/handler.LogoutResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 写入Redis出错
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 用户登出
...
  /register:
    post:
      consumes:
      - multipart/form-data
      description: 使用邮箱、用户名、密码和头像进行注册
      parameters:
      - description: 邮箱地址
        in: formData
        name: email
        required: true
        type: string
      - description: 用户名（2-64字符）
        in: formData
        name: username
        required: true
        type: string
      - description: 密码（8-72字符）
        in: formData
        name: password
        required: true
        type: string
      - description: 确认密码
        in: formData
        name: confirm_password
        required: true
        type: string
      - description: 头像文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功，返回用户信息
          schema:
            $ref: '#/definitions/handler.RegisterResponse'
        "400":
          description: 参数错误或头像上传失败
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: 用户已存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 用户注册
...
  /users/me:
    patch:
      consumes:
      - multipart/form-data
      description: 更新用户邮箱、用户名、密码和头像（头像可选）
      parameters:
      - description: 邮箱地址
        in: formData
        name: email
        type: string
      - description: 用户名（2-64字符）
        in: formData
        name: username
        type: string
      - description: 新密码（8-72字符）
        in: formData
        name: password
        type: string
      - description: 确认新密码
        in: formData
        name: confirm_password
        type: string
      - description: 头像文件（可选）
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功,如更新密码则刷新token
          schema:
            $ref: '#/definitions/handler.UpdateUserResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: 用户名或邮箱已存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新用户信息
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"